Replace internal freetype font handling with VDR's font facilities.
Skins that use fonts other than VDR's standard "Osd", "Fix", and "Sml"
need to be adapted to use font names as understood by VDR's
CreateFont() (Family:Style as of writing this) instead of filenames in
their font attributes' names, to use '@' instead of ':' as the
separator before sizes, and have fonts shipped with the skin installed
and configured in fontconfig.  Typically, install font files (if
necessary) to a dir somewhere, run fc-cache(1) on the dir, replace for
example font="SomeFont.ttf:22,85" with font="Some Font@22,85" or
font="SomeFontBold.ttf:22,85" with font="Some Font:Bold@22,85" in the
*.skin file.  Also, in case the shipped fonts are derivatives of some
existing ones, they may need to be properly renamed to something so
they will be used instead of the original if it happens to be
installed. (closes #36)